Improve map performance by separating HUD render
